Refactor me service
