+++ src/constants/endpoint.ts
@@ -10,10 +10,10 @@ export const ALBUM = {
   getAlbumById: (id: number) => `/albums/${id}?populate=deep,3`
 };
 
-export const PLAYER = {
-  currentSong:
-    '/users/1?populate[song][populate][0]=media&populate[song][populate][1]=artists&populate[album][populate][0]=thumbnail',
-  updateCurrentPlayer: '/users/1'
+export const ME = {
+  info: '/users/1',
+  getCurrentSong:
+    '/users/1?populate[song][populate][0]=media&populate[song][populate][1]=artists&populate[album][populate][0]=thumbnail'
 };
 
 export const COLLECTION = {
