+++ src/services/me.service.ts
@@ -1,0 +1,35 @@
+import { Fetcher } from 'swr';
+
+// Constants
+import { COLLECTION, FetchType, APIKey, ME } from '@constants/index';
+
+// Services
+import { fetcher, PUT, swrFetcher } from './clientRequest';
+
+// Types
+import { GetMyCollection, MeResponse } from '@/types';
+
+// Models
+import { Album, AlbumAttributes } from '@models/index';
+
+export const getCurrentPLayer: Fetcher<MeResponse, APIKey> = () =>
+  swrFetcher(process.env.NEXT_PUBLIC_API_HOST + ME.getCurrentSong);
+
+export const updateCurrentPlayer = (songId: number, albumId: number) =>
+  PUT(ME.info, {
+    song: songId,
+    album: albumId
+  });
+
+export const getMyCollection = async () => {
+  const data = await fetcher<GetMyCollection>(
+    COLLECTION.getMyCollection,
+    FetchType.isr
+  );
+
+  const albums: Album[] = data.albums.map(
+    ({ id, ...rest }) => new Album({ id, attributes: rest as AlbumAttributes })
+  );
+
+  return albums;
+};
